+++ SDL_project_0/Transform.h
@@ -2,18 +2,18 @@
 #include "Vector2.h"
 struct Transform
 {
-	Vector2i position;
+	Vector2d position;
 	double rotation;
 	Vector2f scale;
 	Vector2f pivot = { .5f,.5f }; //only for SDL rotations | probably should move it to RenderObject
 
 	//static const Transform same;
-	Transform(Vector2i pos, double rot, Vector2f sc, Vector2f piv);
-	explicit Transform(Vector2i pos);
+	Transform(Vector2d pos, double rot, Vector2f sc, Vector2f piv);
+	explicit Transform(Vector2d pos);
 	Transform();
-	Transform(Vector2i pos, double rot);
-	Transform(Vector2i pos, double rot, Vector2f sc);
-	Transform(Vector2i pos, Vector2f sc);
+	Transform(Vector2d pos, double rot);
+	Transform(Vector2d pos, double rot, Vector2f sc);
+	Transform(Vector2d pos, Vector2f sc);
 	
 	bool operator ==(const Transform& rhT) {
 		return rhT.position == this->position &&
